move details sidebar into own module
